+++ SR/SR_full_win32.c
@@ -1,6 +1,6 @@
 /**
  *
- *  Copyright (C) 2016-2022 Roman Pauer
+ *  Copyright (C) 2016-2025 Roman Pauer
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy of
  *  this software and associated documentation files (the "Software"), to deal in
@@ -40,7 +40,7 @@ extern const char* ud_reg_tab[];
 int SR_disassemble_remove_segment(const char *ostr, char *dst, uint8_t pfx_seg)
 {
     char *seg, *str1, *str2;
-    int length1;
+    intptr_t length1;
 
     switch (pfx_seg)
     {
@@ -104,7 +104,7 @@ int SR_disassemble_remove_segment(const char *ostr, char *dst, uint8_t pfx_seg)
 int SR_disassemble_remove_segreg2(const char *ostr, char *dst, const struct ud_operand *op)
 {
     char *str1, *str2;
-    int length1, length2;
+    intptr_t length1, length2;
 
     str1 = strchr(ostr, ' ');
     str2 = strchr(ostr, ',');
@@ -140,7 +140,7 @@ int SR_disassemble_remove_segreg2(const char *ostr, char *dst, const struct ud_o
 int SR_disassemble_convert_farcall(const char *ostr, char *dst)
 {
     char *str1, *str2;
-    int length1;
+    intptr_t length1;
 
     str1 = strstr(ostr, " far ");
